% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unchull.R
\name{cart2polar}
\alias{cart2polar}
\title{cart2polar}
\usage{
cart2polar(x, y, xm = median(x), ym = median(y))
}
\arguments{
\item{x}{the x coordinate in Cartesian}

\item{y}{the y coordinate in Cartesian}

\item{xm}{the x coordinate of origin point in polar}

\item{ym}{the y coordinate of origin point in polar}
}
\value{
a list containing r: distance to origin point and t: angle in polar coordinate system
}
\description{
convert Cartesian coordinates to polar coordinates
}
