% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unchull.R
\name{polar2cart}
\alias{polar2cart}
\title{polar2cart}
\usage{
polar2cart(r, t, xm, ym)
}
\arguments{
\item{r}{the distance to origin point in polar}

\item{t}{the angle in polar}

\item{xm}{the x coordinate of origin point in polar}

\item{ym}{the y coordinate of origin point in polar}
}
\value{
a list containing x and y in Cartesian coordinate system
}
\description{
convert polar coordinates to Cartesian coordinates
}
