% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unchull.R
\name{step1_denoise}
\alias{step1_denoise}
\title{step1_denoise}
\usage{
step1_denoise(x, y, n)
}
\arguments{
\item{x}{a vector of x coordinates}

\item{y}{a vector of y coordinates}

\item{n}{parameter for kNN algorithm to defining the point density}
}
\description{
Denoising by keeping only the largest cell cluster using dbscan algorithm.
}
