% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unchull.R
\name{step2_gethull}
\alias{step2_gethull}
\title{step2_gethull}
\usage{
step2_gethull(x, y, alpha)
}
\arguments{
\item{x}{a vector of x coordinates}

\item{y}{a vector of y coordinates}

\item{alpha}{value of alpha}
}
\description{
Get the boundary of given cell cluster using Alpha Shape algorithm.
}
