% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unchull.R
\name{step3_extend}
\alias{step3_extend}
\title{step3_extend}
\usage{
step3_extend(x, y, delta)
}
\arguments{
\item{x}{a vector of x coordinates}

\item{y}{a vector of y coordinates}

\item{delta}{distance to shift the boundary}
}
\description{
Extend the boundary of given cell cluster using polygon-clipping.
}
