% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unchull.R
\name{step4_simplify}
\alias{step4_simplify}
\title{step4_simplify}
\usage{
step4_simplify(x, y, th)
}
\arguments{
\item{x}{a vector of x coordinates}

\item{y}{a vector of y coordinates}

\item{th}{distance to distribute vertexes}
}
\description{
Simplify the boundary of given cell cluster using Polyline Simplification algorithm.
}
