% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unchull.R
\name{step5_smooth}
\alias{step5_smooth}
\title{step5_smooth}
\usage{
step5_smooth(x, y)
}
\arguments{
\item{x}{a vector of x coordinates}

\item{y}{a vector of y coordinates}
}
\description{
Get an X-spline using boundary polygon as control points.
}
